refactor: Move AuthMiddleware to dedicated middleware file
- Create new `middleware.go` in `internal/api/v2/` package
- Migrate existing AuthMiddleware from `system.go` to new file
- Enhance token validation with more robust error handling
- Add support for Bearer token authentication
- Improve error responses for unauthorized access attempts
